Added the new back-end data.  Updated proteome to whatever version was available May 5, 2020.  Did not add the new accessions to the already existing db.  Instead just started fresh.  Due to the fact that there were some errors in the data from phobius, predisi, etc due to the isoform data being used in those data sources but cannonical sequences used for the main data.  Before data went to a text file with no head then was read into an sqlite db.  Now data is put straight away in to the db.
In the future, data should be added to the already existing db rather than redo-ing each time.  This makes it easier because phobius, predisi, etc have limits on the number of sequences that can be uploaded at a time.  I have created a readme to cover that  in the data dir.